requirements.txt
anthropic
googlesearch-python
requests
beautifulsoup4
reportlab
python-dotenv